Socket Initialization on user login
